LPS-167967 Move ObjectFilterUtil to package com.liferay.object.filter.util
