+++ modules/apps/object/object-admin-rest-api/src/main/java/com/liferay/object/admin/rest/dto/v1_0/util/ObjectFieldSettingUtil.java
@@ -17,10 +17,10 @@ package com.liferay.object.admin.rest.dto.v1_0.util;
 import com.liferay.object.admin.rest.dto.v1_0.ObjectStateFlow;
 import com.liferay.object.constants.ObjectFieldConstants;
 import com.liferay.object.constants.ObjectFieldSettingConstants;
+import com.liferay.object.filter.util.ObjectFilterUtil;
 import com.liferay.object.model.ObjectField;
 import com.liferay.object.model.ObjectFieldSetting;
 import com.liferay.object.service.ObjectStateFlowLocalServiceUtil;
-import com.liferay.object.util.ObjectFilterUtil;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
